+++ src/test/kotlin/test/de/itemis/mps/gradle/ModelCheckWithPluginTest.kt
@@ -1,6 +1,7 @@
 package test.de.itemis.mps.gradle
 
 import de.itemis.mps.gradle.ErrorMessages
+import de.itemis.mps.gradle.modelcheck.ModelCheckPluginExtensions
 import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.TaskOutcome
 import org.hamcrest.CoreMatchers
@@ -273,8 +274,9 @@ class ModelCheckWithPluginTest {
             .withPluginClasspath()
             .buildAndFail()
 
-        MatcherAssert.assertThat(result.output, CoreMatchers.containsString(ErrorMessages.MUST_SET_VERSION_AND_LOCATION))
+        MatcherAssert.assertThat(result.output, CoreMatchers.containsString(ErrorMessages.mustSetConfigOrLocation("modelcheck")))
     }
+
     @Test
     fun `check model fails with only MPS path set`() {
         settingsFile.writeText(settingsBoilerplate())
@@ -296,6 +298,7 @@ class ModelCheckWithPluginTest {
             .withPluginClasspath()
             .buildAndFail()
 
-        MatcherAssert.assertThat(result.output, CoreMatchers.containsString(ErrorMessages.MUST_SET_CONFIG_OR_VERSION))
+        MatcherAssert.assertThat(result.output, CoreMatchers.containsString(
+            ErrorMessages.mustSetVersionWhenNoMpsConfiguration("modelcheck")))
     }
 }